+++ addon/components/mobile/object-list-view.js
@@ -112,7 +112,7 @@ export default ObjectListViewComponent.extend({
       let allSelect = this.get('allSelect');
 
       let rootItem = {
-        icon: 'icon-guideline-check-menu icon',
+        icon: 'dropdown icon',
         iconAlignment: 'right',
         title: '',
         items: [],
@@ -189,12 +189,11 @@ export default ObjectListViewComponent.extend({
       if (this.get('allSelect')) {
         this.send('checkAll');
       } else {
-        let selectedRecords = this.get('selectedRecords');
-        for (let i = 0; i < selectedRecords.length; i++) {
-          this.send('selectRow', selectedRecords[i], {
-            checked: false
-          });
-        }
+        const contentWithKeys = this.get('contentWithKeys');
+        this.get('selectedRecords').map((record) => contentWithKeys.findBy('data', record)).forEach((modelWithKey) => {
+          modelWithKey.set('selected', false);
+          this.send('selectRow', modelWithKey, { checked: false });
+        });
       }
     },
   }